admin cli: consumer-groups => groups
