+++ kafka/cli/admin/__init__.py
@@ -7,7 +7,7 @@ from kafka.admin.client import KafkaAdminClient
 from .acls import ACLsSubCommand
 from .cluster import ClusterSubCommand
 from .configs import ConfigsSubCommand
-from .consumer_groups import ConsumerGroupsSubCommand
+from .groups import GroupsSubCommand
 from .partitions import PartitionsSubCommand
 from .topics import TopicsSubCommand
 from .users import UsersSubCommand
@@ -51,7 +51,7 @@ def run_cli(args=None):
     subparsers = parser.add_subparsers(help='subcommands')
     for cmd in [ACLsSubCommand, ClusterSubCommand, ConfigsSubCommand,
                 TopicsSubCommand, PartitionsSubCommand,
-                ConsumerGroupsSubCommand, UsersSubCommand]:
+                GroupsSubCommand, UsersSubCommand]:
         cmd.add_subparser(subparsers)
     config = parser.parse_args(args)
