+++ kafka/cli/admin/groups/list_offsets.py
@@ -1,7 +1,7 @@
-class ListConsumerGroupOffsets:
+class ListGroupOffsets:
 
     @classmethod
     def add_subparser(cls, subparsers):
-        parser = subparsers.add_parser('list-offsets', help='List Offsets for Consumer Group')
+        parser = subparsers.add_parser('list-offsets', help='List Offsets for Group')
         parser.add_argument('-g', '--group-id', type=str, required=True)
         parser.set_defaults(command=lambda cli, args: cli.list_consumer_group_offsets(args.group_id))
